requirements.txt
flask
flask-sqlalchemy
werkzeug
pillow
qrcode
opencv-python